# docs/swagger.yaml
basePath: /
definitions:
  models.Book:
    properties:
      authors:
        items:
          type: string
        type: array
      categories:
        items:
          type: string
        type: array
      id:
        type: integer
      isbn:
        type: string
      longDescription:
        type: string
      pageCount:
        type: integer
      publishedDate:
        type: string
      shortDescription:
        type: string
      status:
        type: string
      thumbnailUrl:
        type: string
      title:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    name: Garrett Omi
  description: This is the finest book application ever made
  title: byFood Book Application
  version: "1.0"
paths:
  /books:
    get:
      description: Get all books from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a new book in the database
      parameters:
      - description: Book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    get:
      description: Get a book by its ID from the database
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
      summary: Get book by ID
      tags:
      - books
swagger: "2.0"